requirements/local.txt
-r ./base.txt

# Debugging
ipdb==0.11

# Tools
django-extensions==2.2.9

# Email
django-anymail[mailgun]==5.0

# Testing
mypy==0.650
pytest==6.2.2
pytest-sugar==0.9.4
pytest-django==4.1.0
factory-boy==3.2.0